Remove Reread from op status
Addressing mode functions now perform a reread automatically on cycle 5
and cycle 6 for absolute indexed and indirect indexed modes,
respectively. This doesn't mess with read instructions, because they
will be done before then if a page boundary isn't crossed. If a page
boundary is crossed, a reread needs to happen anyway. As for write
instructions, the read-in value will be thrown away regardless.