Add PPU, refactor CPU and other classes
Implemented a (mostly) pixel-accurate PPU that successfully displays the
background for mapper 0 games, such as Donkey Kong, Super Mario Bros., and
Pac-Man. MMC was also updated to support mapper 0 games.

Refactored interrupt polling a bit. Polling is now performed on the
second-to-last cycle rather than the last cycle, which is how it is done on the
NES.

CPU power-up state is more accurate now by starting with the stack pointer as
0xfd and the interrupt disable flag set in the status register. This required
updating the unit tests.

Renamed the Op class to CPUOp now that PPUOp exists.

Renamed driver.cpp to emulator.cpp.

Made member functions const and private wherever possible.